Rename tabs and declaration technique, fix e2es
