+++ content/webapp/components/Tabs/README.md
@@ -2,7 +2,7 @@
 To navigate through related sections within a set of results.
 
 ### Default
-`Tabs.Default` uses real tab behaviour where clicking a tab will swap out the content of the tabpanel underneath.
+`Tabs.Tab` uses real tab behaviour where clicking a tab will swap out the content of the tabpanel underneath.
 
 ### Links
-`Tabs.Links` uses the same styled-components as `Tabs.Default`, but the tabs function as anchors which link to a new page.
+`Tabs.Anchor` uses the same styled-components as `Tabs.Tab`, but the tabs function as anchors which link to a new page.
