--- playwright/test/selectors/concepts.ts
@@ -1,3 +1,0 @@
-export const worksAboutThisPerson = 'css=button#tab-worksAbout';
-
-export const imagesAboutThisPerson = 'css=button#tab-imagesAbout';
